fix: slider view
- Move slider styles to a separate file
- Add GlobalStyles usage to MediaScreen component
